requirements.txt
Flask==2.3.2
Pillow==10.0.0
easyocr==1.6.2
torch==2.0.1
numpy==1.24.4